Fixes minor issues, stated in the PR review
